requirements.txt
sentence-transformers
faiss-cpu
transformers
accelerate
bitsandbytes
torch
